fix: some versions of RN not working with Collections
