+++ .changeset/young-bulldogs-scream.md
@@ -1,0 +1,8 @@
+---
+'@data-client/endpoint': patch
+'@rest-hooks/endpoint': patch
+---
+
+Fix some versions of RN not working with Collections
+
+(Array.sort() does not exist in all versions)
